feat: weight 타입 double로 수정

+++ src/main/java/Asset/Thing.java
@@ -1,7 +1,7 @@
 package Asset;
 
 public interface Thing {
-    int getWeight();
+    double getWeight();
 
-    void setWeight(int weight);
+    void setWeight(double weight);
 }
